requirements.txt
pandas
numpy
scikit-learn
pymongo
requests
python-dotenv
pytz
xgboost